Validation now injected for whole assembly!

+++ InternetShopAPI/Controllers/Requests/ProductaddQuantityRequest.cs
@@ -1,5 +1,4 @@
 ﻿namespace InternetShopAPI.Controllers.Requests;
 
 public sealed record ProductaddQuantityRequest(
-    ushort Quantity
-    );
+    int Quantity);
